requirements.txt
streamlit
pandas
openpyxl
gtts
groq
pdfplumber
PyMuPDF
python-docx
tqdm
numpy
pillow
regex
